Amiibo Service live tests refactoring (#15)
This PR contains the work done to refactor the test cases for the `AmiiboService` service type in the test target, to write them as the `Swift Testing` framework intended. In. addition, the documentation tasks in the `Makefile` file and the generated `DocC` documentation for Github Pages have also been updated.

Reviewed-on: https://repo.rock-n-code.com/rock-n-code/amiibo-service/pulls/15
Co-authored-by: Javier Cicchelli <javier@rock-n-code.com>
Co-committed-by: Javier Cicchelli <javier@rock-n-code.com>